src/requirements.txt
rpyc==4.1.2

cloudshell-orch-core>=2.0.0.0,<2.2.0.0
cloudshell-shell-core>=4,<5
cloudshell-automation-api>=9.1,<9.2
cloudshell-traffic==3.0.2

pybyteblower==0.1
